Update Trend_Finder
Update: Version 1.2
In this update, I have added a new variable show_all.

If you set this variable to true, the program will first retrieve the list of all symbols and then analyze all the symbols. However, please note that this process takes a significant amount of time as it needs to download a large amount of data.